requirements.txt
# PDF Processing
pdfplumber>=0.10.0
pypdf>=3.0.0

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Database - PostgreSQL
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Vector Search (pgvector)
pgvector>=0.2.0

# OpenAI (embeddings)
openai>=1.0.0

# Anthropic (AI features)
anthropic>=0.18.0

# Data processing
pandas>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Configuration
python-dotenv>=1.0.0

# HTTP client for scraping/monitoring
httpx>=0.26.0
aiohttp>=3.9.0

# Web Scraping (Bulletin)
playwright>=1.40.0
beautifulsoup4>=4.12.0
firecrawl-py>=1.0.0

# Scheduling for URL monitoring
apscheduler>=3.10.0

# Payments (Stripe)
stripe>=7.0.0

# Development/Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
anthropic>=0.76.0
